// clients/pricingService/models/PriceListCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PriceListCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    name: string;
    description?: string | null;
    startDate?: string;
    endDate?: string;
    currencyId?: string | null;
    unitId?: string | null;
    unitGroupId?: string | null;
    tenantId?: string | null;
    enrollmentId?: string | null;
};